requirements.txt
pyyaml
mysql-connector
gunicorn
django
django_extensions
psycopg2
python-magic
